update: 代码优化、增加延时队列
