+++ common-utils-core/src/main/java/com/el/common/constant/MailConfigConstants.java
@@ -7,17 +7,12 @@ public class MailConfigConstants {
 
     public static final String SSL_SOCKET_FACTORY_CLASS = "mail.imap.socketFactory.class";
     public static final String SSL_SOCKET_FACTORY = "javax.net.ssl.SSLSocketFactory";
-
     public static final String MAIL_HOST_KEY = "mail.imap.host";
     public static final String MAIL_HOST_VALUE = "smtp.misumi.com.cn";
-
     public static final String SSL_SOCKET_FACTORY_FALLBACK = "mail.imap.socketFactory.fallback";
-
     public static final String MAIL_IMAP_PORT_KEY = "mail.imap.port";
-
     public static final String MAIL_IMAP_AUTH_KEY = "mail.imap.auth";
     public static final String SSL_SOCKET_FACTORY_PORT_KEY = "mail.imap.socketFactory.port";
-
     public static final String MAIL_IMAP_PORT_VALUE = "993";
 
 }
