--- common-utils-core/src/main/java/com/el/common/mail/SendMailService.java
@@ -1,12 +1,0 @@
-package com.el.common.mail;
-
-/**
- * since 8/16/21
- *
- * @author eddie
- */
-public interface SendMailService {
-
-
-
-}
