+++ common-utils-core/src/main/java/com/el/common/mail/utils/MailUtils.java
@@ -1,0 +1,259 @@
+package com.el.common.mail.utils;
+
+import com.el.common.mail.data.MailAttachment;
+import org.apache.commons.lang3.StringUtils;
+
+import javax.mail.*;
+import javax.mail.internet.InternetAddress;
+import javax.mail.internet.MimeMessage;
+import javax.mail.internet.MimeMultipart;
+import javax.mail.internet.MimeUtility;
+import java.io.IOException;
+import java.io.UnsupportedEncodingException;
+import java.text.SimpleDateFormat;
+import java.util.Date;
+
+/**
+ * 邮件工具类
+ * since 6/15/22
+ *
+ * @author eddie
+ */
+public class MailUtils {
+
+    /**
+     * 获得邮件主题
+     *
+     * @param msg 邮件内容
+     * @return 解码后的邮件主题
+     */
+    public static String getSubject(MimeMessage msg) throws UnsupportedEncodingException, MessagingException {
+        return MimeUtility.decodeText(msg.getSubject());
+    }
+
+    /**
+     * 获得邮件发件人
+     *
+     * @param msg 邮件内容
+     * @return 姓名 <Email地址>
+     */
+    public static String getFrom(MimeMessage msg) throws MessagingException, UnsupportedEncodingException {
+        Address[] froms = msg.getFrom();
+        if (froms.length < 1) {
+            throw new MessagingException("没有发件人!");
+        }
+
+        InternetAddress address = (InternetAddress) froms[0];
+        String person = address.getPersonal();
+        if (person != null) {
+            person = MimeUtility.decodeText(person) + " ";
+        } else {
+            person = "";
+        }
+        String from = person + "<" + address.getAddress() + ">";
+
+        return address.getAddress();
+    }
+
+    /**
+     * 根据收件人类型，获取邮件收件人、抄送和密送地址。如果收件人类型为空，则获得所有的收件人
+     * <p>Message.RecipientType.TO  收件人</p>
+     * <p>Message.RecipientType.CC  抄送</p>
+     * <p>Message.RecipientType.BCC 密送</p>
+     *
+     * @param msg  邮件内容
+     * @param type 收件人类型
+     * @return 收件人1 <邮件地址1>, 收件人2 <邮件地址2>
+     */
+    public static String getReceiveAddress(MimeMessage msg, Message.RecipientType type) throws MessagingException {
+        StringBuilder receiveAddress = new StringBuilder();
+        Address[] addressArray;
+        if (type == null) {
+            addressArray = msg.getAllRecipients();
+        } else {
+            addressArray = msg.getRecipients(type);
+        }
+        if (addressArray == null || addressArray.length < 1) {
+            throw new MessagingException("没有收件人!");
+        }
+        for (Address address : addressArray) {
+            InternetAddress internetAddress = (InternetAddress) address;
+            receiveAddress.append(internetAddress.toUnicodeString()).append(",");
+        }
+        receiveAddress.deleteCharAt(receiveAddress.length() - 1);
+        return receiveAddress.toString();
+    }
+
+    /**
+     * 获得邮件发送时间
+     *
+     * @param msg 邮件内容
+     * @return yyyy年mm月dd日 星期X HH:mm
+     */
+    public static String getSentDate(MimeMessage msg, String pattern) throws MessagingException {
+        Date receivedDate = msg.getSentDate();
+        if (receivedDate == null) {
+            return "";
+        }
+        if (pattern == null || "".equals(pattern)) {
+            pattern = "yyyy年MM月dd日 E HH:mm ";
+        }
+        return new SimpleDateFormat(pattern).format(receivedDate);
+    }
+
+    /**
+     * 判断邮件中是否包含附件
+     *
+     * @param part 邮件内容
+     * @return 邮件中存在附件返回true，不存在返回false
+     */
+    public static boolean isContainAttachment(Part part) throws MessagingException, IOException {
+        boolean flag = false;
+        if (part.isMimeType("multipart/*")) {
+            MimeMultipart multipart = (MimeMultipart) part.getContent();
+            int partCount = multipart.getCount();
+            for (int i = 0; i < partCount; i++) {
+                BodyPart bodyPart = multipart.getBodyPart(i);
+                String disp = bodyPart.getDisposition();
+                if (disp != null && (disp.equalsIgnoreCase(Part.ATTACHMENT) || disp.equalsIgnoreCase(Part.INLINE))) {
+                    flag = true;
+                } else if (bodyPart.isMimeType("multipart/*")) {
+                    flag = isContainAttachment(bodyPart);
+                } else {
+                    String contentType = bodyPart.getContentType();
+                    if (contentType.contains("application")) {
+                        flag = true;
+                    }
+
+                    if (contentType.contains("name")) {
+                        flag = true;
+                    }
+                }
+                if (flag) {
+                    break;
+                }
+            }
+        } else if (part.isMimeType("message/rfc822")) {
+            flag = isContainAttachment((Part) part.getContent());
+        }
+        return flag;
+    }
+
+    /**
+     * 判断邮件是否已读
+     *
+     * @param msg 邮件内容
+     * @return 如果邮件已读返回true, 否则返回false
+     */
+    public static boolean isSeen(MimeMessage msg) throws MessagingException {
+        return msg.getFlags().contains(Flags.Flag.SEEN);
+    }
+
+    /**
+     * 判断邮件是否需要阅读回执
+     *
+     * @param msg 邮件内容
+     * @return 需要回执返回true, 否则返回false
+     */
+    public static boolean isReplySign(MimeMessage msg) throws MessagingException {
+        boolean replySign = false;
+        String[] headers = msg.getHeader("Disposition-Notification-To");
+        if (headers != null) {
+            replySign = true;
+        }
+        return replySign;
+    }
+
+    /**
+     * 获得邮件的优先级
+     *
+     * @param msg 邮件内容
+     * @return 1(High):紧急  3:普通(Normal)  5:低(Low)
+     */
+    public static String getPriority(MimeMessage msg) throws MessagingException {
+        String priority = "普通";
+        String[] headers = msg.getHeader("X-Priority");
+        if (headers != null) {
+            String headerPriority = headers[0];
+            if (headerPriority.contains("1") || headerPriority.contains("High")) {
+                return "紧急";
+            } else if (headerPriority.contains("5") || headerPriority.contains("Low")) {
+                return "低";
+            } else {
+                return "普通";
+            }
+        }
+        return priority;
+    }
+
+    /**
+     * 获得邮件文本内容
+     *
+     * @param part    邮件体
+     * @param content 存储邮件文本内容的字符串
+     */
+    public static void getMailTextContent(Part part, StringBuffer content) throws MessagingException, IOException {
+        //如果是文本类型的附件，通过getContent方法可以取到文本内容，但这不是我们需要的结果，所以在这里要做判断
+        boolean isContainTextAttach = part.getContentType().indexOf("name") > 0;
+        if (part.isMimeType("text/*") && !isContainTextAttach) {
+            content.append(part.getContent().toString());
+        } else if (part.isMimeType("message/rfc822")) {
+            getMailTextContent((Part) part.getContent(), content);
+        } else if (part.isMimeType("multipart/*")) {
+            Multipart multipart = (Multipart) part.getContent();
+            int partCount = multipart.getCount();
+            for (int i = 0; i < partCount; i++) {
+                BodyPart bodyPart = multipart.getBodyPart(i);
+                getMailTextContent(bodyPart, content);
+            }
+        }
+    }
+
+    public static MailAttachment getFileInputStream(Part part) throws Exception {
+        String fileName;
+        MailAttachment mailAttachment = new MailAttachment();
+        if (part.isMimeType("multipart/*")) {
+            Multipart mp = (Multipart) part.getContent();
+            for (int i = 0; i < mp.getCount(); i++) {
+                BodyPart mPart = mp.getBodyPart(i);
+                String disposition = mPart.getDisposition();
+                if (StringUtils.isBlank(disposition)) {
+                    continue;
+                }
+                if (disposition.equals(Part.ATTACHMENT) || disposition.equals(Part.INLINE)) {
+                    fileName = decodeText(mPart.getFileName());
+                    mailAttachment.setFileName(fileName);
+                    mailAttachment.setAttachment(mPart.getInputStream());
+                    return mailAttachment;
+                } else if (mPart.isMimeType("multipart/*")) {
+                    return getFileInputStream(mPart);
+                } else {
+                    fileName = mPart.getFileName();
+                    if ((fileName != null)) {
+                        fileName = decodeText(fileName);
+                        mailAttachment.setFileName(fileName);
+                        mailAttachment.setAttachment(mPart.getInputStream());
+                        return mailAttachment;
+                    }
+                }
+            }
+        } else if (part.isMimeType("message/rfc822")) {
+            return getFileInputStream((Part) part.getContent());
+        }
+        return mailAttachment;
+    }
+
+    /**
+     * 文本解码
+     *
+     * @param encodeText 解码MimeUtility.encodeText(String text)方法编码后的文本
+     * @return 解码后的文本
+     */
+    public static String decodeText(String encodeText) throws UnsupportedEncodingException {
+        if (encodeText == null || "".equals(encodeText)) {
+            return "";
+        } else {
+            return MimeUtility.decodeText(encodeText);
+        }
+    }
+}
