+++ common-utils-core/src/main/java/com/el/common/time/entity/DateEntity.java
@@ -14,8 +14,12 @@ import lombok.NoArgsConstructor;
 @NoArgsConstructor
 @AllArgsConstructor
 public class DateEntity {
-
+    /**
+     * 开始时间
+     */
     private String startTime;
-
+    /**
+     * 结束时间
+     */
     private String endTime;
 }
